% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structure.R
\name{create_strucutre}
\alias{create_strucutre}
\title{A function to establish a minimum folder structure}
\usage{
create_strucutre(
  project_name = NULL,
  project_description = NULL,
  author = NULL,
  git_ignore = FALSE,
  ignore = c("r", "macos")
)
}
\arguments{
\item{project_name}{Name of the Project (character string)}

\item{project_description}{Description of the project (character string or vector for multiple authors)}

\item{author}{Author/s (character string or vector for multiple authors)}

\item{git_ignore}{Include git.ignore file. Defaults to FALSE. If true character string must be passed to ignore}

\item{ignore}{programs to create gitignore file for. Defaulsts to c("r", "macos").}
}
\value{

}
\description{
This function will create a code and data folder within the project working directory
The code folder will contain 3 folders: helper_functions, data_cleaning, analysis
The data folder will contain two folders: raw, processed
There is also the option to add a git.ignore set up for macOS and R
}
\examples{
create_strucutre(git_ignore = TRUE, ignore = c("macos", "r")
}
